+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Write the date in place of the "Unreleased" in the case a new version is release
 
 ## Unreleased
 
+- Add client_secret,redirect_on_success,redirect_on_failure to ProxiedOIDCAuthenticator. This is to allow login using Tiled-UI
+- Allow configuration of user_id_claim for OIDCAuthenticator
+
 ### Fixed
 
 - Fix the webhook `history` and `delete` endpoints when a catalog is mounted under